update rust to 1.77.2

+++ shared/simulator/tests/api_test.rs
@@ -77,7 +77,7 @@ fn test_id() {
                 vector![0.0, 0.0],
                 vector![0.0, 0.0],
                 0.0,
-                fighter(team as i32),
+                fighter(team),
             )
         })
         .collect::<Vec<_>>();
